Added build system and modularize codebase
