+++ go.mod
@@ -1,0 +1,3 @@
+module macup
+
+go 1.24.3
